pyproject.toml
[tool.poetry]
name = "accession-number-looker-upper"
version = "0.1.0"
description = ""
authors = ["Anthony Shvets <anthonyshvets@gmail.com>"]
readme = "README.md"
packages = [{include = "accession_number_looker_upper"}]

[tool.poetry.dependencies]
python = "^3.11"
biopython = "^1.81"
typer = "^0.7.0"
pandas = "^1.5.3"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
